CFB Recruiting Information - Position Groups

If only start_year is provided, function will get CFB recruiting information based on position groups during that year for all FBS teams.

cfb_recruiting_position(
  start_year = NULL,
  end_year = NULL,
  team = NULL,
  conference = NULL
)

Arguments

start_year
(Integer optional): Start Year, 4 digit format (YYYY). Note: 2000 is the minimum value

end_year
(Integer optional): End Year, 4 digit format (YYYY). Note: 2020 is the maximum value currently

team
(String optional): Team - Select a valid team, D-I football

conference
(String optional): Conference name - select a valid FBS conference Conference names P5: ACC, Big 12, Big Ten, SEC, Pac-12 Conference names G5 and FBS Independents: Conference USA, Mid-American, Mountain West, FBS Independents, American Athletic

Details

If you would like CFB recruiting information for players, please see the cfbrecruitR::cfb_recruiting_player() function

If you would like CFB recruiting information for teams, please see the cfbrecruitR::cfb_recruiting_team() function

Examples

cfb_recruiting_position(2018, team="Texas")

#>    team conference  positionGroup averageRating totalRating commits
#> 1 Texas     Big 12 Defensive Back     0.9480923     12.3252      13
#> 2 Texas     Big 12 Defensive Line     0.9050300      9.0503      10
#> 3 Texas     Big 12     Linebacker     0.9040250      7.2322       8
#> 4 Texas     Big 12 Offensive Line     0.9099364     10.0093      11
#> 5 Texas     Big 12    Quarterback     0.9323400      4.6617       5
#> 6 Texas     Big 12       Receiver     0.9238545     10.1624      11
#> 7 Texas     Big 12   Running Back     0.9463667      2.8391       3
#> 8 Texas     Big 12  Special Teams     0.8816556      7.9349       9
#> 9 Texas     Big 12  All Positions     0.9173586     64.2151      70
#>   averageStars
#> 1     4.076923
#> 2     3.500000
#> 3     3.500000
#> 4     3.909091
#> 5     4.000000
#> 6     3.727273
#> 7     4.333333
#> 8     3.555556
#> 9     3.785714

cfb_recruiting_position(2016, 2020, team="Virginia")

#>       team conference  positionGroup averageRating totalRating commits
#> 1 Virginia        ACC Defensive Back     0.8332929     11.6661      14
#> 2 Virginia        ACC Defensive Line     0.8512125     13.6194      16
#> 3 Virginia        ACC     Linebacker     0.8460889     15.2296      18
#> 4 Virginia        ACC Offensive Line     0.8452263     16.0593      19
#> 5 Virginia        ACC    Quarterback     0.8465500      5.0793       6
#> 6 Virginia        ACC       Receiver     0.8401950     16.8039      20
#> 7 Virginia        ACC   Running Back     0.8453714      5.9176       7
#> 8 Virginia        ACC  Special Teams     0.8361111      7.5250       9
#> 9 Virginia        ACC  All Positions     0.8431211     91.9002     109
#>   averageStars
#> 1     2.857143
#> 2     3.125000
#> 3     3.000000
#> 4     3.052632
#> 5     3.000000
#> 6     2.950000
#> 7     3.000000
#> 8     2.888889
#> 9     2.990826

cfb_recruiting_position(2015, 2020, conference = "SEC")

#>                  team conference  positionGroup averageRating totalRating
#> 1             Alabama        SEC Defensive Back     0.9465250     22.7166
#> 2             Alabama        SEC Defensive Line     0.9543161     29.5838
#> 3             Alabama        SEC     Linebacker     0.9390412     15.9637
#> 4             Alabama        SEC Offensive Line     0.9327783     21.4539
#> 5             Alabama        SEC    Quarterback     0.9061444      8.1553
#> 6             Alabama        SEC       Receiver     0.9246696     21.2674
#> 7             Alabama        SEC   Running Back     0.9298571     13.0180
#> 8             Alabama        SEC  Special Teams     0.8863800      8.8638
#> 9            Arkansas        SEC Defensive Back     0.8627654     22.4319
#> 10           Arkansas        SEC Defensive Line     0.8787962     22.8487
#> 11           Arkansas        SEC     Linebacker     0.8500211     16.1504
#> 12           Arkansas        SEC Offensive Line     0.8516045     18.7353
#> 13           Arkansas        SEC    Quarterback     0.8731125      6.9849
#> 14           Arkansas        SEC       Receiver     0.8798105     16.7164
#> 15           Arkansas        SEC   Running Back     0.8774571      6.1422
#> 16           Arkansas        SEC  Special Teams     0.8589000      8.5890
#> 17             Auburn        SEC Defensive Back     0.8867667     15.9618
#> 18             Auburn        SEC Defensive Line     0.9239870     21.2517
#> 19             Auburn        SEC     Linebacker     0.9038818     19.8854
#> 20             Auburn        SEC Offensive Line     0.8983118     15.2713
#> 21             Auburn        SEC    Quarterback     0.8898800      4.4494
#> 22             Auburn        SEC       Receiver     0.9005708     21.6137
#> 23             Auburn        SEC   Running Back     0.9129800      9.1298
#> 24             Auburn        SEC  Special Teams     0.8934955     19.6569
#> 25            Florida        SEC Defensive Back     0.9020100     27.0603
#> 26            Florida        SEC Defensive Line     0.9040182     19.8884
#> 27            Florida        SEC     Linebacker     0.8898467     13.3477
#> 28            Florida        SEC Offensive Line     0.8827667     23.8347
#> 29            Florida        SEC    Quarterback     0.8857143      6.2000
#> 30            Florida        SEC       Receiver     0.9096542     21.8317
#> 31            Florida        SEC   Running Back     0.8969375      7.1755
#> 32            Florida        SEC  Special Teams     0.8576000      4.2880
#> 33            Georgia        SEC Defensive Back     0.9102037     24.5755
#> 34            Georgia        SEC Defensive Line     0.9288633     27.8659
#> 35            Georgia        SEC     Linebacker     0.9426188     15.0819
#> 36            Georgia        SEC Offensive Line     0.9216385     23.9626
#> 37            Georgia        SEC    Quarterback     0.9209375      7.3675
#> 38            Georgia        SEC       Receiver     0.9206640     23.0166
#> 39            Georgia        SEC   Running Back     0.9420625      7.5365
#> 40            Georgia        SEC  Special Teams     0.8812250     10.5747
#> 41           Kentucky        SEC Defensive Back     0.8647150     17.2943
#> 42           Kentucky        SEC Defensive Line     0.8649654     22.4891
#> 43           Kentucky        SEC     Linebacker     0.8635083     10.3621
#> 44           Kentucky        SEC Offensive Line     0.8633545     18.9938
#> 45           Kentucky        SEC    Quarterback     0.8507000      5.9549
#> 46           Kentucky        SEC       Receiver     0.8583818     18.8844
#> 47           Kentucky        SEC   Running Back     0.8686875      6.9495
#> 48           Kentucky        SEC  Special Teams     0.8524353     14.4914
#> 49                LSU        SEC Defensive Back     0.9374136     20.6231
#> 50                LSU        SEC Defensive Line     0.9160400     27.4812
#> 51                LSU        SEC     Linebacker     0.9289917     11.1479
#> 52                LSU        SEC Offensive Line     0.9100680     22.7517
#> 53                LSU        SEC    Quarterback     0.8728000      6.9824
#> 54                LSU        SEC       Receiver     0.9143160     22.8579
#> 55                LSU        SEC   Running Back     0.9257546     10.1833
#> 56                LSU        SEC  Special Teams     0.8538308     11.0998
#> 57  Mississippi State        SEC Defensive Back     0.8741955     19.2323
#> 58  Mississippi State        SEC Defensive Line     0.8818545     19.4008
#> 59  Mississippi State        SEC     Linebacker     0.9073500      7.2588
#> 60  Mississippi State        SEC Offensive Line     0.8599889     15.4798
#> 61  Mississippi State        SEC    Quarterback     0.8763167      5.2579
#> 62  Mississippi State        SEC       Receiver     0.8560118     14.5522
#> 63  Mississippi State        SEC   Running Back     0.8826429      6.1785
#> 64  Mississippi State        SEC  Special Teams     0.8604706     14.6280
#> 65           Missouri        SEC Defensive Back     0.8484773     18.6665
#> 66           Missouri        SEC Defensive Line     0.8598200     12.8973
#> 67           Missouri        SEC     Linebacker     0.8507400     12.7611
#> 68           Missouri        SEC Offensive Line     0.8567450     17.1349
#> 69           Missouri        SEC    Quarterback     0.8807000      4.4035
#> 70           Missouri        SEC       Receiver     0.8580182     18.8764
#> 71           Missouri        SEC   Running Back     0.8418875      6.7351
#> 72           Missouri        SEC  Special Teams     0.8499800      8.4998
#> 73           Ole Miss        SEC Defensive Back     0.8617200     17.2344
#> 74           Ole Miss        SEC Defensive Line     0.8743650     17.4873
#> 75           Ole Miss        SEC     Linebacker     0.8567067     12.8506
#> 76           Ole Miss        SEC Offensive Line     0.8755840     21.8896
#> 77           Ole Miss        SEC    Quarterback     0.8837667      7.9539
#> 78           Ole Miss        SEC       Receiver     0.9119895     17.3278
#> 79           Ole Miss        SEC   Running Back     0.8900900      8.9009
#> 80           Ole Miss        SEC  Special Teams     0.8872000      3.5488
#> 81     South Carolina        SEC Defensive Back     0.8699762     18.2695
#> 82     South Carolina        SEC Defensive Line     0.8877385     23.0812
#> 83     South Carolina        SEC     Linebacker     0.8669000     12.1366
#> 84     South Carolina        SEC Offensive Line     0.8646154     22.4800
#> 85     South Carolina        SEC    Quarterback     0.9170571      6.4194
#> 86     South Carolina        SEC       Receiver     0.8708458     20.9003
#> 87     South Carolina        SEC   Running Back     0.8800222      7.9202
#> 88     South Carolina        SEC  Special Teams     0.8664364      9.5308
#> 89          Tennessee        SEC Defensive Back     0.8831739     20.3130
#> 90          Tennessee        SEC Defensive Line     0.9003042     21.6073
#> 91          Tennessee        SEC     Linebacker     0.8987000     14.3792
#> 92          Tennessee        SEC Offensive Line     0.8976727     19.7488
#> 93          Tennessee        SEC    Quarterback     0.8861889      7.9757
#> 94          Tennessee        SEC       Receiver     0.8761043     20.1504
#> 95          Tennessee        SEC   Running Back     0.8638636      9.5025
#> 96          Tennessee        SEC  Special Teams     0.8588333     12.8825
#> 97          Texas A&M        SEC Defensive Back     0.9180450     18.3609
#> 98          Texas A&M        SEC Defensive Line     0.9079800     22.6995
#> 99          Texas A&M        SEC     Linebacker     0.8815600     17.6312
#> 100         Texas A&M        SEC Offensive Line     0.8876538     23.0790
#> 101         Texas A&M        SEC    Quarterback     0.8954875      7.1639
#> 102         Texas A&M        SEC       Receiver     0.9124037     24.6349
#> 103         Texas A&M        SEC   Running Back     0.8864333     10.6372
#> 104         Texas A&M        SEC  Special Teams     0.8424000      8.4240
#> 105        Vanderbilt        SEC Defensive Back     0.8578895     16.2999
#> 106        Vanderbilt        SEC Defensive Line     0.8307143     11.6300
#> 107        Vanderbilt        SEC     Linebacker     0.8476769     11.0198
#> 108        Vanderbilt        SEC Offensive Line     0.8469783     19.4805
#> 109        Vanderbilt        SEC    Quarterback     0.8761167      5.2567
#> 110        Vanderbilt        SEC       Receiver     0.8486955     18.6713
#> 111        Vanderbilt        SEC   Running Back     0.8454625      6.7637
#> 112        Vanderbilt        SEC  Special Teams     0.8228400      8.2284
#> 113           Alabama        SEC  All Positions     0.9339238    141.0230
#> 114          Arkansas        SEC  All Positions     0.8656847    118.5990
#> 115            Auburn        SEC  All Positions     0.9022695    127.2200
#> 116           Florida        SEC  All Positions     0.8958428    123.6260
#> 117           Georgia        SEC  All Positions     0.9209289    139.9810
#> 118          Kentucky        SEC  All Positions     0.8613396    115.4190
#> 119               LSU        SEC  All Positions     0.9118308    133.1270
#> 120 Mississippi State        SEC  All Positions     0.8716949    101.9880
#> 121          Missouri        SEC  All Positions     0.8544838     99.9746
#> 122          Ole Miss        SEC  All Positions     0.8786336    107.1930
#> 123    South Carolina        SEC  All Positions     0.8749130    120.7380
#> 124         Tennessee        SEC  All Positions     0.8850308    126.5590
#> 125         Texas A&M        SEC  All Positions     0.8961527    132.6310
#> 126        Vanderbilt        SEC  All Positions     0.8465243     97.3503
#>     commits averageStars
#> 1        24     4.000000
#> 2        31     4.193548
#> 3        17     4.000000
#> 4        23     3.956522
#> 5         9     3.888889
#> 6        23     3.826087
#> 7        14     3.928571
#> 8        10     3.500000
#> 9        26     3.230769
#> 10       26     3.307692
#> 11       19     3.052632
#> 12       22     3.045455
#> 13        8     3.500000
#> 14       19     3.368421
#> 15        7     3.142857
#> 16       10     3.200000
#> 17       18     3.500000
#> 18       23     3.782609
#> 19       22     3.727273
#> 20       17     3.470588
#> 21        5     3.600000
#> 22       24     3.541667
#> 23       10     3.900000
#> 24       22     3.590909
#> 25       30     3.500000
#> 26       22     3.818182
#> 27       15     3.400000
#> 28       27     3.444444
#> 29        7     3.428571
#> 30       24     3.583333
#> 31        8     3.500000
#> 32        5     3.000000
#> 33       27     3.666667
#> 34       30     3.966667
#> 35       16     4.062500
#> 36       26     3.769231
#> 37        8     4.000000
#> 38       25     3.840000
#> 39        8     4.125000
#> 40       12     3.583333
#> 41       20     3.100000
#> 42       26     3.192308
#> 43       12     3.333333
#> 44       22     3.181818
#> 45        7     3.000000
#> 46       22     3.090909
#> 47        8     3.125000
#> 48       17     3.176471
#> 49       22     4.000000
#> 50       30     3.666667
#> 51       12     3.916667
#> 52       25     3.600000
#> 53        8     3.375000
#> 54       25     3.800000
#> 55       11     3.909091
#> 56       13     3.076923
#> 57       22     3.136364
#> 58       22     3.409091
#> 59        8     3.375000
#> 60       18     3.111111
#> 61        6     3.500000
#> 62       17     3.058824
#> 63        7     3.428571
#> 64       17     3.000000
#> 65       22     3.045455
#> 66       15     3.200000
#> 67       15     3.000000
#> 68       20     3.150000
#> 69        5     3.400000
#> 70       22     3.045455
#> 71        8     3.000000
#> 72       10     2.900000
#> 73       20     3.050000
#> 74       20     3.250000
#> 75       15     3.066667
#> 76       25     3.200000
#> 77        9     3.444444
#> 78       19     3.684211
#> 79       10     3.500000
#> 80        4     3.250000
#> 81       21     3.142857
#> 82       26     3.500000
#> 83       14     3.214286
#> 84       26     3.115385
#> 85        7     3.714286
#> 86       24     3.208333
#> 87        9     3.111111
#> 88       11     3.363636
#> 89       23     3.304348
#> 90       24     3.583333
#> 91       16     3.562500
#> 92       22     3.545455
#> 93        9     3.444444
#> 94       23     3.260870
#> 95       11     3.181818
#> 96       15     3.066667
#> 97       20     3.750000
#> 98       25     3.640000
#> 99       20     3.300000
#> 100      26     3.461538
#> 101       8     3.500000
#> 102      27     3.666667
#> 103      12     3.416667
#> 104      10     3.000000
#> 105      19     3.105263
#> 106      14     2.785714
#> 107      13     3.000000
#> 108      23     3.000000
#> 109       6     3.166667
#> 110      22     3.090909
#> 111       8     3.000000
#> 112      10     2.700000
#> 113     151     3.960265
#> 114     137     3.218978
#> 115     141     3.631206
#> 116     138     3.521739
#> 117     152     3.848684
#> 118     134     3.156716
#> 119     146     3.698630
#> 120     117     3.205128
#> 121     117     3.076923
#> 122     122     3.286885
#> 123     138     3.268116
#> 124     143     3.384615
#> 125     148     3.513514
#> 126     115     2.991304